+++ app/src/test/java/woowacourse/shopping/CartProductTest.kt
@@ -1,76 +1,24 @@
 package woowacourse.shopping
 
-import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
-import woowacourse.shopping.domain.CartProducts
+import org.junit.jupiter.api.assertDoesNotThrow
+import org.junit.jupiter.api.assertThrows
+import woowacourse.shopping.domain.CartProduct
 import woowacourse.shopping.domain.Product
 
 class CartProductTest {
     @Test
-    fun `상품을 추가하면 해당 상품이 포함된다`() {
-        val cartProducts1 = CartProducts()
-        val newProduct =
-            Product(
-                imageUri = "image",
-                name = "twohander",
-                price = 10000,
-            )
-        val cartProducts2 = cartProducts1.add(newProduct)
-
-        assertTrue(cartProducts2.items.contains(newProduct))
+    fun `개수는 0 이상이라면 예외가 발생한다`() {
+        assertThrows <IllegalArgumentException> {
+            CartProduct(Product(imageUri = "image", name = "name", price = 10000), -1)
+        }
     }
 
     @Test
-    fun `상품을 제거하면 해당 상품이 포함되지 않는다`() {
-        val newProduct =
-            Product(
-                imageUri = "image",
-                name = "twohander",
-                price = 10000,
-            )
-        val cartProducts1 = CartProducts(products = listOf(newProduct))
-        val targetId = newProduct.uuid
-        val cartProducts2 = cartProducts1.remove(targetId)
-
-        assertTrue(cartProducts2.items.contains(newProduct).not())
-    }
-
-    @Test
-    fun `상품을 id로 검색 한다`() {
-        val newProduct =
-            Product(
-                imageUri = "image",
-                name = "twohander",
-                price = 10000,
-            )
-        val cartProducts1 = CartProducts(products = listOf(newProduct))
-        val targetId = newProduct.uuid
-
-        val foundProduct = cartProducts1.findWithId(targetId)
-
-        assertEquals(newProduct, foundProduct)
-    }
-
-    @Test
-    fun `id 검색에 실패했다면 null을 반환한다`() {
-        val newProduct =
-            Product(
-                imageUri = "image",
-                name = "twohander",
-                price = 10000,
-            )
-        val newProduct2 =
-            Product(
-                imageUri = "image",
-                name = "samuel",
-                price = 50,
-            )
-        val cartProducts1 = CartProducts(products = listOf(newProduct))
-        val targetId = newProduct2.uuid
-
-        val foundProduct = cartProducts1.findWithId(targetId)
-
-        assertEquals(null, foundProduct)
+    fun `개수가 0 이상이라면 예외가 발생하지 않는다`() {
+        assertDoesNotThrow {
+            CartProduct(Product(imageUri = "image", name = "name", price = 10000), 0)
+        }
     }
 }
